+++ include/peer/concurrency_control/deterministic/write_based/wb_coordinator.h
@@ -9,21 +9,8 @@
 
 namespace peer::cc {
     class WBCoordinator : public Coordinator<WBWorkerFSM, WBReserveTable, WBCoordinator> {
-    public:
-        // NOT thread safe
-        // processTxnList will take the transactions and move them back after execution
-        bool processTxnList(std::vector<std::unique_ptr<proto::Transaction>>& txnList) {
-            int totalWorkerCount = (int)workerList.size();
+        bool processSync(const auto& afterStart, const auto& afterCommit) {
             reserveTable->reset();
-            // prepare txn function
-            auto afterStart = [&](const Worker& worker, WBWorkerFSM& fsm) {
-                auto& fsmTxnList = fsm.getMutableTxnList();
-                fsmTxnList.clear();
-                auto id = worker.getId();
-                for (int i = id; i < (int)txnList.size(); i += totalWorkerCount) {
-                    fsmTxnList.push_back(std::move(txnList[i]));
-                }
-            };
             auto ret = processParallel(InvokerCommand::START, ReceiverState::READY, afterStart);
             if (!ret) {
                 LOG(ERROR) << "init txnList failed!";
@@ -40,14 +27,6 @@ namespace peer::cc {
                 LOG(ERROR) << "first commit failed!";
                 return false;
             }
-            // move back
-            auto afterCommit = [&](const Worker& worker, WBWorkerFSM& fsm) {
-                auto& fsmTxnList = fsm.getMutableTxnList();
-                auto id = worker.getId();
-                for (int i = id, j = 0; i < (int)txnList.size(); i += totalWorkerCount) {
-                    txnList[i] = std::move(fsmTxnList[j++]);
-                }
-            };
             // Second commit
             ret = processParallel(InvokerCommand::COMMIT, ReceiverState::FINISH_COMMIT, afterCommit);
             if (!ret) {
